Player and first bg sprite
Implemented a placeholder player sprite + code to make it chnage directions based on the button inputs. Have attempted to include a bg sprite that moves in a circle with cos+sin functions, however, they seem to have a small error meaning the sprite cannot move in a perfect circle and as such does not stay lined up.