Jumping error
Started building a level but quickly ran into an issue.
I never limited jumping and trying to rectify that has led to an error.
Although Unity is correctly stopping the player from falling through the platforms it is not detecting the collision.
As such OnCollisionEnter2D cannot see the "Floor" tag and cannot switch the player out of a jumping state.
Tl:Dr - the player can either jump once or infinitely.